requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.28.0
azure-storage-blob==12.17.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
isodate==0.6.1
joblib==1.3.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
nltk==3.8.1
numpy==1.25.1
openai==0.27.8
pandas==2.0.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
regex==2023.8.8
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
